+++ RenderComponent/StructuredBuffer.cpp
@@ -9,7 +9,7 @@ StructuredBuffer::StructuredBuffer(
 ) : elements(elementsCount), offsets(elementsCount)
 {
 	memcpy(elements.data(), elementsArray, sizeof(StructuredBufferElement) * elementsCount);
-	size_t offst = 0;
+	uint64_t offst = 0;
 	for (UINT i = 0; i < elementsCount; ++i)
 	{
 		offsets[i] = offst;
@@ -35,11 +35,11 @@ void StructuredBuffer::TransformBufferState(ID3D12GraphicsCommandList* commandLi
 	));
 }
 
-size_t StructuredBuffer::GetStride(UINT index) const
+uint64_t StructuredBuffer::GetStride(UINT index) const
 {
 	return elements[index].stride;
 }
-size_t StructuredBuffer::GetElementCount(UINT index) const
+uint64_t StructuredBuffer::GetElementCount(UINT index) const
 {
 	return elements[index].elementCount;
 }
@@ -64,7 +64,7 @@ D3D12_GPU_VIRTUAL_ADDRESS StructuredBuffer::GetAddress(UINT element, UINT index)
 #endif
 }
 
-size_t StructuredBuffer::GetAddressOffset(UINT element, UINT index) const
+uint64_t StructuredBuffer::GetAddressOffset(UINT element, UINT index) const
 {
 #ifdef NDEBUG
 	auto& ele = elements[element];
